+++ src/renderer/views/functions/manage-category.vue
