+++ src/renderer/views/functions/menu.vue
@@ -53,8 +53,3 @@ export default {
   }
 }
 </script>
-
-<style lang="less">
-#menu-functions {
-}
-</style>
