+++ src/renderer/views/functions/search-result-item.vue
